// packages/mdx-kit/src/types/gatsby-graphql-types.ts
export type Maybe<T> = T | null
// ...
export type Date = any

/** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
export type Json = any

// ====================================================
// Scalars
// ====================================================

// ====================================================
// Interfaces
// ====================================================

/** Node Interface */
export interface Node {
  id: string

  parent?: Maybe<Node>

  children: Node[]

  internal: Internal
}

// ====================================================
// Types
// ====================================================

export interface Query {
  file?: Maybe<File>

  allFile: FileConnection

  mdx?: Maybe<Mdx>

  allMdx: MdxConnection

  imageSharp?: Maybe<ImageSharp>

  allImageSharp: ImageSharpConnection

  sitePage?: Maybe<SitePage>

  allSitePage: SitePageConnection

  sitePlugin?: Maybe<SitePlugin>

  allSitePlugin: SitePluginConnection

  site?: Maybe<Site>

  allSite: SiteConnection

  directory?: Maybe<Directory>

  allDirectory: DirectoryConnection

  authorJson?: Maybe<AuthorJson>

  allAuthorJson: AuthorJsonConnection

  svg?: Maybe<Svg>

  allSvg: SvgConnection
}

export interface File extends Node {
  birthtime?: Maybe<Date>

  birthtimeMs?: Maybe<number>

  sourceInstanceName?: Maybe<string>

  absolutePath?: Maybe<string>

  relativePath?: Maybe<string>

  extension?: Maybe<string>

  size?: Maybe<number>

  prettySize?: Maybe<string>

  modifiedTime?: Maybe<Date>

  accessTime?: Maybe<Date>

  changeTime?: Maybe<Date>

  birthTime?: Maybe<Date>

  root?: Maybe<string>

  dir?: Maybe<string>

  base?: Maybe<string>

  ext?: Maybe<string>

  name?: Maybe<string>

  relativeDirectory?: Maybe<string>

  dev?: Maybe<number>

  mode?: Maybe<number>

  nlink?: Maybe<number>

  uid?: Maybe<number>

  gid?: Maybe<number>

  rdev?: Maybe<number>

  blksize?: Maybe<number>

  ino?: Maybe<number>

  blocks?: Maybe<number>

  atimeMs?: Maybe<number>

  mtimeMs?: Maybe<number>

  ctimeMs?: Maybe<number>

  atime?: Maybe<Date>

  mtime?: Maybe<Date>

  ctime?: Maybe<Date>
  /** Copy file to static directory and return public url to it */
  publicURL?: Maybe<string>

  childImageSharp?: Maybe<ImageSharp>

  id: string

  parent?: Maybe<Node>

  children: Node[]

  internal: Internal

  childMdx?: Maybe<Mdx>

  childAuthorJson?: Maybe<AuthorJson>

  childSvg?: Maybe<Svg>
}

export interface Internal {
  content?: Maybe<string>

  contentDigest: string

  description?: Maybe<string>

  fieldOwners?: Maybe<(Maybe<string>)[]>

  ignoreType?: Maybe<boolean>

  mediaType?: Maybe<string>

  owner: string

  type: string
}

export interface ImageSharp extends Node {
  fixed?: Maybe<ImageSharpFixed>

  resolutions?: Maybe<ImageSharpResolutions>

  fluid?: Maybe<ImageSharpFluid>

  sizes?: Maybe<ImageSharpSizes>

  original?: Maybe<ImageSharpOriginal>

  resize?: Maybe<ImageSharpResize>

  id: string

  parent?: Maybe<Node>

  children: Node[]

  internal: Internal
}

export interface ImageSharpFixed {
  base64?: Maybe<string>

  tracedSVG?: Maybe<string>

  aspectRatio?: Maybe<number>

  width?: Maybe<number>

  height?: Maybe<number>

  src?: Maybe<string>

  srcSet?: Maybe<string>

  srcWebp?: Maybe<string>

  srcSetWebp?: Maybe<string>

  originalName?: Maybe<string>
}

export interface ImageSharpResolutions {
  base64?: Maybe<string>

  tracedSVG?: Maybe<string>

  aspectRatio?: Maybe<number>

  width?: Maybe<number>

  height?: Maybe<number>

  src?: Maybe<string>

  srcSet?: Maybe<string>

  srcWebp?: Maybe<string>

  srcSetWebp?: Maybe<string>

  originalName?: Maybe<string>
}

export interface ImageSharpFluid {
  base64?: Maybe<string>

  tracedSVG?: Maybe<string>

  aspectRatio?: Maybe<number>

  src?: Maybe<string>

  srcSet?: Maybe<string>

  srcWebp?: Maybe<string>

  srcSetWebp?: Maybe<string>

  sizes?: Maybe<string>

  originalImg?: Maybe<string>

  originalName?: Maybe<string>

  presentationWidth?: Maybe<number>

  presentationHeight?: Maybe<number>
}

export interface ImageSharpSizes {
  base64?: Maybe<string>

  tracedSVG?: Maybe<string>

  aspectRatio?: Maybe<number>

  src?: Maybe<string>

  srcSet?: Maybe<string>

  srcWebp?: Maybe<string>

  srcSetWebp?: Maybe<string>

  sizes?: Maybe<string>

  originalImg?: Maybe<string>

  originalName?: Maybe<string>

  presentationWidth?: Maybe<number>

  presentationHeight?: Maybe<number>
}

export interface ImageSharpOriginal {
  width?: Maybe<number>

  height?: Maybe<number>

  src?: Maybe<string>
}

export interface ImageSharpResize {
  src?: Maybe<string>

  tracedSVG?: Maybe<string>

  width?: Maybe<number>

  height?: Maybe<number>

  aspectRatio?: Maybe<number>

  originalName?: Maybe<string>
}

export interface Mdx extends Node {
  rawBody: string

  fileAbsolutePath: string

  frontmatter?: Maybe<MdxFrontmatter>

  body: string

  excerpt: string

  headings?: Maybe<(Maybe<MdxHeadingMdx>)[]>

  html?: Maybe<string>

  mdxAST?: Maybe<Json>

  tableOfContents?: Maybe<Json>

  timeToRead?: Maybe<number>

  wordCount?: Maybe<MdxWordCount>

  fields?: Maybe<MdxFields>

  id: string

  parent?: Maybe<Node>

  children: Node[]

  internal: Internal
}

export interface MdxFrontmatter {
  title: string

  route?: Maybe<boolean>

  version?: Maybe<number>

  author?: Maybe<string>

  slug?: Maybe<string>

  date?: Maybe<Date>
}

export interface MdxHeadingMdx {
  value?: Maybe<string>

  depth?: Maybe<number>
}

export interface MdxWordCount {
  paragraphs?: Maybe<number>

  sentences?: Maybe<number>

  words?: Maybe<number>
}

export interface MdxFields {
  slug?: Maybe<string>

  sourceInstanceName?: Maybe<string>

  relativePath?: Maybe<string>

  uid?: Maybe<string>

  title?: Maybe<string>

  route?: Maybe<boolean>

  author?: Maybe<AuthorJson>

  date?: Maybe<Date>

  category?: Maybe<string>
}

export interface AuthorJson extends Node {
  id: string

  parent?: Maybe<Node>

  children: Node[]

  internal: Internal

  bio?: Maybe<string>

  avatar?: Maybe<File>

  twitter?: Maybe<string>

  github?: Maybe<string>
}

export interface Svg extends Node {
  id: string

  parent?: Maybe<Node>

  children: Node[]

  internal: Internal

  content?: Maybe<string>

  path?: Maybe<string>

  sourceInstanceName?: Maybe<string>
}

export interface FileConnection {
  totalCount: number

  edges: FileEdge[]

  nodes: File[]

  pageInfo: PageInfo

  distinct: string[]

  group: FileGroupConnection[]
}

export interface FileEdge {
  next?: Maybe<File>

  node: File

  previous?: Maybe<File>
}

export interface PageInfo {
  currentPage: number

  hasPreviousPage: boolean

  hasNextPage: boolean

  itemCount: number

  pageCount: number

  perPage?: Maybe<number>
}

export interface FileGroupConnection {
  totalCount: number

  edges: FileEdge[]

  nodes: File[]

  pageInfo: PageInfo

  field: string

  fieldValue?: Maybe<string>
}

export interface MdxConnection {
  totalCount: number

  edges: MdxEdge[]

  nodes: Mdx[]

  pageInfo: PageInfo

  distinct: string[]

  group: MdxGroupConnection[]
}

export interface MdxEdge {
  next?: Maybe<Mdx>

  node: Mdx

  previous?: Maybe<Mdx>
}

export interface MdxGroupConnection {
  totalCount: number

  edges: MdxEdge[]

  nodes: Mdx[]

  pageInfo: PageInfo

  field: string

  fieldValue?: Maybe<string>
}

export interface ImageSharpConnection {
  totalCount: number

  edges: ImageSharpEdge[]

  nodes: ImageSharp[]

  pageInfo: PageInfo

  distinct: string[]

  group: ImageSharpGroupConnection[]
}

export interface ImageSharpEdge {
  next?: Maybe<ImageSharp>

  node: ImageSharp

  previous?: Maybe<ImageSharp>
}

export interface ImageSharpGroupConnection {
  totalCount: number

  edges: ImageSharpEdge[]

  nodes: ImageSharp[]

  pageInfo: PageInfo

  field: string

  fieldValue?: Maybe<string>
}

export interface SitePage extends Node {
  id: string

  parent?: Maybe<Node>

  children: Node[]

  internal: Internal

  path?: Maybe<string>

  internalComponentName?: Maybe<string>

  component?: Maybe<string>

  componentChunkName?: Maybe<string>

  isCreatedByStatefulCreatePages?: Maybe<boolean>

  context?: Maybe<SitePageContext>

  pluginCreator?: Maybe<SitePlugin>

  pluginCreatorId?: Maybe<string>

  componentPath?: Maybe<string>
}

export interface SitePageContext {
  frontmatter?: Maybe<SitePageContextFrontmatter>
}

export interface SitePageContextFrontmatter {
  title?: Maybe<string>

  route?: Maybe<boolean>

  version?: Maybe<number>

  author?: Maybe<string>

  slug?: Maybe<string>

  date?: Maybe<Date>
}

export interface SitePlugin extends Node {
  id: string

  parent?: Maybe<Node>

  children: Node[]

  internal: Internal

  resolve?: Maybe<string>

  name?: Maybe<string>

  version?: Maybe<string>

  pluginOptions?: Maybe<SitePluginPluginOptions>

  nodeAPIs?: Maybe<(Maybe<string>)[]>

  browserAPIs?: Maybe<(Maybe<string>)[]>

  ssrAPIs?: Maybe<(Maybe<string>)[]>

  pluginFilepath?: Maybe<string>

  packageJson?: Maybe<SitePluginPackageJson>
}

export interface SitePluginPluginOptions {
  path?: Maybe<string>

  name?: Maybe<string>

  extensions?: Maybe<(Maybe<string>)[]>

  defaultLayouts?: Maybe<SitePluginPluginOptionsDefaultLayouts>

  gatsbyRemarkPlugins?: Maybe<(Maybe<SitePluginPluginOptionsGatsbyRemarkPlugins>)[]>

  short_name?: Maybe<string>

  start_url?: Maybe<string>

  background_color?: Maybe<string>

  theme_color?: Maybe<string>

  display?: Maybe<string>

  icon?: Maybe<string>

  logo?: Maybe<string>

  appName?: Maybe<string>

  appDescription?: Maybe<string>

  developerName?: Maybe<string>

  dir?: Maybe<string>

  lang?: Maybe<string>

  background?: Maybe<string>

  orientation?: Maybe<string>

  version?: Maybe<string>

  icons?: Maybe<SitePluginPluginOptionsIcons>

  modules?: Maybe<(Maybe<string>)[]>

  pathCheck?: Maybe<boolean>
}
// ...
export interface SitePageConnection {
  totalCount: number

  edges: SitePageEdge[]

  nodes: SitePage[]

  pageInfo: PageInfo

  distinct: string[]

  group: SitePageGroupConnection[]
}

export interface SitePageEdge {
  next?: Maybe<SitePage>

  node: SitePage

  previous?: Maybe<SitePage>
}

export interface SitePageGroupConnection {
  totalCount: number

  edges: SitePageEdge[]

  nodes: SitePage[]

  pageInfo: PageInfo

  field: string

  fieldValue?: Maybe<string>
}

export interface SitePluginConnection {
  totalCount: number

  edges: SitePluginEdge[]

  nodes: SitePlugin[]

  pageInfo: PageInfo

  distinct: string[]

  group: SitePluginGroupConnection[]
}

export interface SitePluginEdge {
  next?: Maybe<SitePlugin>

  node: SitePlugin

  previous?: Maybe<SitePlugin>
}

export interface SitePluginGroupConnection {
  totalCount: number

  edges: SitePluginEdge[]

  nodes: SitePlugin[]

  pageInfo: PageInfo

  field: string

  fieldValue?: Maybe<string>
}

export interface Site extends Node {
  id: string

  parent?: Maybe<Node>

  children: Node[]

  internal: Internal

  siteMetadata?: Maybe<SiteSiteMetadata>

  port?: Maybe<number>

  host?: Maybe<string>

  pathPrefix?: Maybe<string>

  mapping?: Maybe<SiteMapping>

  polyfill?: Maybe<boolean>

  buildTime?: Maybe<Date>
}

export interface SiteSiteMetadata {
  title?: Maybe<string>

  author?: Maybe<string>

  description?: Maybe<string>

  siteUrl?: Maybe<string>

  org?: Maybe<string>

  contact?: Maybe<string>

  favicon?: Maybe<string>

  faviconSvg?: Maybe<string>

  logo?: Maybe<string>

  icon?: Maybe<string>
}

export interface SiteMapping {
  Mdx_fields_author?: Maybe<string>
}

export interface SiteConnection {
  totalCount: number

  edges: SiteEdge[]

  nodes: Site[]

  pageInfo: PageInfo

  distinct: string[]

  group: SiteGroupConnection[]
}

export interface SiteEdge {
  next?: Maybe<Site>

  node: Site

  previous?: Maybe<Site>
}

export interface SiteGroupConnection {
  totalCount: number

  edges: SiteEdge[]

  nodes: Site[]

  pageInfo: PageInfo

  field: string

  fieldValue?: Maybe<string>
}

export interface Directory extends Node {
  id: string

  parent?: Maybe<Node>

  children: Node[]

  internal: Internal

  sourceInstanceName?: Maybe<string>

  absolutePath?: Maybe<string>

  relativePath?: Maybe<string>

  extension?: Maybe<string>

  size?: Maybe<number>

  prettySize?: Maybe<string>

  modifiedTime?: Maybe<Date>

  accessTime?: Maybe<Date>

  changeTime?: Maybe<Date>

  birthTime?: Maybe<Date>

  root?: Maybe<string>

  dir?: Maybe<string>

  base?: Maybe<string>

  ext?: Maybe<string>

  name?: Maybe<string>

  relativeDirectory?: Maybe<string>

  dev?: Maybe<number>

  mode?: Maybe<number>

  nlink?: Maybe<number>

  uid?: Maybe<number>

  gid?: Maybe<number>

  rdev?: Maybe<number>

  blksize?: Maybe<number>

  ino?: Maybe<number>

  blocks?: Maybe<number>

  atimeMs?: Maybe<number>

  mtimeMs?: Maybe<number>

  ctimeMs?: Maybe<number>

  birthtimeMs?: Maybe<number>

  atime?: Maybe<Date>

  mtime?: Maybe<Date>

  ctime?: Maybe<Date>

  birthtime?: Maybe<Date>
}

export interface DirectoryConnection {
  totalCount: number

  edges: DirectoryEdge[]

  nodes: Directory[]

  pageInfo: PageInfo

  distinct: string[]

  group: DirectoryGroupConnection[]
}

export interface DirectoryEdge {
  next?: Maybe<Directory>

  node: Directory

  previous?: Maybe<Directory>
}

export interface DirectoryGroupConnection {
  totalCount: number

  edges: DirectoryEdge[]

  nodes: Directory[]

  pageInfo: PageInfo

  field: string

  fieldValue?: Maybe<string>
}

export interface AuthorJsonConnection {
  totalCount: number

  edges: AuthorJsonEdge[]

  nodes: AuthorJson[]

  pageInfo: PageInfo

  distinct: string[]

  group: AuthorJsonGroupConnection[]
}

export interface AuthorJsonEdge {
  next?: Maybe<AuthorJson>

  node: AuthorJson

  previous?: Maybe<AuthorJson>
}

export interface AuthorJsonGroupConnection {
  totalCount: number

  edges: AuthorJsonEdge[]

  nodes: AuthorJson[]

  pageInfo: PageInfo

  field: string

  fieldValue?: Maybe<string>
}

export interface SvgConnection {
  totalCount: number

  edges: SvgEdge[]

  nodes: Svg[]

  pageInfo: PageInfo

  distinct: string[]

  group: SvgGroupConnection[]
}